FedClusterEnsemble/requirements.txt
aiohappyeyeballs==2.6.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
aiohttp-cors==0.8.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
aiohttp==3.11.14 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
aiosignal==1.3.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
ansicon==1.89.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0" and platform_system == "Windows"
antlr4-python3-runtime==4.9.3 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
async-timeout==5.0.1 ; python_full_version >= "3.10.0" and python_version < "3.11"
attrs==25.3.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
blessed==1.20.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
cachetools==5.5.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
certifi==2025.1.31 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
cffi==1.17.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
charset-normalizer==3.4.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
click==8.1.8 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
colorama==0.4.6 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0" and platform_system == "Windows"
colorful==0.5.6 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
cryptography==41.0.7 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
distlib==0.3.9 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
filelock==3.18.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
flwr[simulation]==1.5.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
frozenlist==1.5.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
google-api-core==2.24.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
google-auth==2.38.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
googleapis-common-protos==1.69.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
gpustat==1.1.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
grpcio==1.71.0 ; python_version >= "3.10" and python_full_version < "3.11.0"
hydra-core==1.3.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
idna==3.10 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
iterators==0.0.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
jinxed==1.3.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0" and platform_system == "Windows"
jsonschema-specifications==2024.10.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
jsonschema==4.23.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
msgpack==1.1.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
multidict==6.2.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
numpy==1.26.4 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
nvidia-ml-py==12.570.86 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
omegaconf==2.3.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
opencensus-context==0.1.3 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
opencensus==0.11.4 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
packaging==24.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
pillow==11.1.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
platformdirs==3.11.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
prometheus-client==0.21.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
propcache==0.3.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
proto-plus==1.26.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
protobuf==3.20.3 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
psutil==7.0.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
py-spy==0.4.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
pyasn1-modules==0.4.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
pyasn1==0.6.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
pycparser==2.22 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
pycryptodome==3.22.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
pydantic==1.10.21 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
pyyaml==6.0.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
ray[default]==2.6.3 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
referencing==0.36.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
requests==2.32.3 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
rpds-py==0.23.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
rsa==4.9 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
six==1.17.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
smart-open==7.1.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
torch==1.12.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
torchvision==0.13.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
tqdm==4.66.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
typing-extensions==4.12.2 ; python_full_version >= "3.10.0" and python_version < "3.11"
urllib3==2.3.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
virtualenv==20.21.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
wcwidth==0.2.13 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
wrapt==1.12.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
yarl==1.18.3 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
